+++ collective/mcp/HISTORY.rst
@@ -1,7 +1,14 @@
 Changelog
 =========
 
-1.0dev (unreleased)
+0.2 (unreleased)
+----------------
+
+ - Added possibility to set custom CSS class to the subpage. By
+   default, it has the class 'mcp_widget_XXX' where XXX is the widget
+   id. [vincent]
+
+0.1  (25/02/2011)
 -------------------
 
 - Initial release